Skip files if they've already been symlinked to the correct place
Update Rakefile so that if the target_path is already a symlink to
the source_path we just skip it and don't prompt about overwriting.

+++ Rakefile
@@ -15,6 +15,10 @@ def ignore_files
   [ '.DS_Store', '.git', '.gitignore', '.', '..' ]
 end
 
+def already_symlinked?(source_path, target_path)
+  File.exists?(target_path) && File.lstat(target_path).symlink? && File.readlink(target_path) == source_path
+end
+
 def create_symlinks
   dir = File.dirname(__FILE__)
 
@@ -22,20 +26,22 @@ def create_symlinks
     homedir = File.expand_path ENV['HOME']
     source_path = File.join dir, file
     target_path = File.join homedir, file
-    if File.exists?(target_path)
-      puts "File #{target_path} exists.  Overwrite it (y/n)?"
-      if STDIN.gets.chomp.downcase == 'y'
-        puts "#{"DELETING".red} #{target_path}"
-        raise "This shouldn't happen, but if it does, I'm refusing to delete /" if target_path == "/"
-        FileUtils.rm_r(target_path)
-      else
-        puts "#{"SKIPPING".blue} #{target_path}"
-        next
+    unless already_symlinked(source_path, target_path)
+      if File.exists?(target_path)
+        puts "File #{target_path} exists.  Overwrite it (y/n)?"
+        if STDIN.gets.chomp.downcase == 'y'
+          puts "#{"DELETING".red} #{target_path}"
+          raise "This shouldn't happen, but if it does, I'm refusing to delete /" if target_path == "/"
+          FileUtils.rm_r(target_path)
+        else
+          puts "#{"SKIPPING".blue} #{target_path}"
+          next
+        end
       end
-    end
 
-    puts "#{"[SYMLINK]".green} #{source_path} --> #{target_path.yellow}"
-    FileUtils.ln_s(source_path, target_path)
+      puts "#{"[SYMLINK]".green} #{source_path} --> #{target_path.yellow}"
+      FileUtils.ln_s(source_path, target_path)
+    end
   end
 
   puts
